Fixing a bug in parsing integers.
If the character parsing splits the text into two blocks right in the middle of
an integer then we have to put the two halves of the integer back together
to parse its value.  Normally this was done correctly.  But there was a bug if
the second half of the integer started with one or more zeros.

Something like this was happening:  "1234" + "005" --> "12345".


git-svn-id: http://svn.code.sf.net/p/genoviz/code/trunk@84 80d5a9c2-7d8c-4c9f-9369-4e03bca4d2bd