Added the ability to process a new directive "##IGB-hierarchy" that allows simpler processing of GFF files from the WTA chips.
Also fixed a long-standing issue that the use_container flag was only obeyed
if USE_GROUPING was true.  Now the use_container flag is always obeyed.

A simple use of the new flag would be
##IGB-hierarchy  0 transcript  1 exon

That means that "transcript" is a top-level annotation containing one or more  "exon".
You could equivalently  write this on several lines, if that is more convenient
##IGB-hierarchy  0 transcript
##IGB-hierarchy  1 exon


git-svn-id: http://svn.code.sf.net/p/genoviz/code/trunk@125 80d5a9c2-7d8c-4c9f-9369-4e03bca4d2bd