igb_default_prefs.xml:
Adding rule for linking to gene model pages at TAIR

Changing URL for linking to NCBI pages.  Now operating 'get info' function will open the RefSeq (or other mRNA record) page directly, thus saving a mouse click.

WebBrowserControl.java:

Testing for Mac OS X platform.  Using 'open' to invoke browser when this is not already set by the user. Renovating the code somewhat -- now we use a 'switch' statement instead of an if/else to pick the system command for opening a browser. Note: needs to be tested on Linux and Windows tomorrow.


git-svn-id: http://svn.code.sf.net/p/genoviz/code/trunk@197 80d5a9c2-7d8c-4c9f-9369-4e03bca4d2bd