Added a method refreshGenomeSelector() that can re-populate the GUI Genome Selection list when new genomes (AnnotatedSeqGroups) have been added to the GenometryModel.
This is mostly useful when the LoadFileAction has created a fake genome,
called something like "Unknown Genome 1", when loading a file.

There isn't actually any event type to let the QuickLoaderView know when
AnnotatedSeqGroups have been added, so this is actually repopulating
the list every time it receives a SeqGroupChanged event.


git-svn-id: http://svn.code.sf.net/p/genoviz/code/trunk@261 80d5a9c2-7d8c-4c9f-9369-4e03bca4d2bd