This is a mouse listener for the SeqMapView.
This is based on code that was in the SeqMapView class itself.
Now in a new class to make both classes easier to read.

This is not identical to the code from SeqMapView.  The biggest difference
is that now dragging does not automatically select sequence residues rather
than annotations.  To select sequence, you must now start the drag on top of the
axis tier.  Dragging from any other location will select annotations.


git-svn-id: http://svn.code.sf.net/p/genoviz/code/trunk@394 80d5a9c2-7d8c-4c9f-9369-4e03bca4d2bd