Deleted the selected_syms variable.
Instead, call getSelectedSyms() when you need to know what the current
selections are.

The reason is that keeping selected_syms up-to-date required code in
lots of places, and it was occasionally still wrong.
(For instance, if select() was called with call_listeners == false)
Also setSelectedRegion was adding the selected region symmetry to
the selected_syms list, but then this setting was getting erased
in postSelections().

These problems could have been fixed, but it seems easier to get
rid of the variable altogether.

If optimization is needed, it might be possible to use
gmodel.getSelectedSymmetries() in place of calls to getSelectedSyms().


git-svn-id: http://svn.code.sf.net/p/genoviz/code/trunk@446 80d5a9c2-7d8c-4c9f-9369-4e03bca4d2bd