Got rid of the variables last_selected_glyph and last_selected_sym.  There were cases when these contained incorrect values.  Besides, it is more intuitive to care about the complete set of what is selected, not just the last thing that was added to the selections.
One thing that is improved by throwing out these variables
is the setAnnotatedSeq() method when it is called with
the same seq and preserve_selections = true.  Now we
preserve (as much as possible) the entire set of what was
selected, not just one single selection out of many.

Putting text in the status bar to tell when an
item is selected.

Made copySelectedResidues() behave more intuitively.
If there is a selected sequence region, use that as the
basis of the residues that get copied to the clipboard.
Else if there is one (and only one) selected glyph, use
that.  Else pop-up an error.
In any case where something is copied to the clipboard,
describe it in the status bar.


git-svn-id: http://svn.code.sf.net/p/genoviz/code/trunk@450 80d5a9c2-7d8c-4c9f-9369-4e03bca4d2bd