Several refinements to the drag-selection behavior:
Dragging will now skip any items in the axis tier (like contigs).  This allows you to drag
to include items above and below the axis without selecting a contig.  (You rarely want
to select a contig, and doing so can lead to excess memory consumption in the sliced
view)

Drags starting in the axis tier will now select ONLY sequence residues.  Will not
also select the item that is under the mouse at the point where the mouse is
released.


git-svn-id: http://svn.code.sf.net/p/genoviz/code/trunk@498 80d5a9c2-7d8c-4c9f-9369-4e03bca4d2bd