Implemented IPrefEditorComponent so this can be used as a preferences panel.
Re-arranged columns in the table, and removed the "Show" column.
(You can manipulate hiding and showing things in other ways, and this
data is not persistent.)

Made "Glyph Depth" display as a boolean, since there are only two valid choices,
corresponding to "connected" or "not connected".

Made it impossible to edit some properties of the "default" style.


git-svn-id: http://svn.code.sf.net/p/genoviz/code/trunk@633 80d5a9c2-7d8c-4c9f-9369-4e03bca4d2bd