Added some protection against the problem where a server may return two multipart annotations (a group, or transcript) with the same group id in different tiers (different method type names).
Basically, I replaced a hashtable of group id's with a hash of hashtables, such
that there is a different hashtable for each tier.

This prevents the parser from getting confused by two items having the same ID,
and this is a small improvment over the current case.

However, if FILTER_OUT_BY_ID is true, which it always is, then only one of the
items with the same ID will be loaded, because the global symhash does not
hash things in different tiers, or even in different genomes, separately.


git-svn-id: http://svn.code.sf.net/p/genoviz/code/trunk@757 80d5a9c2-7d8c-4c9f-9369-4e03bca4d2bd