Modified the behavior in two ways:
1) If getInputStream() is called without specifying the cache usage pattern, then it now
takes the user's setting for the prefered cache usage rather than assuming
NORMAL_USAGE.  This allows the program to be used in off-line mode, since if
the user has specified ONLY_CACHE, it will now not need to access the internet
to load the default synonyms file.

2) Now getInputStream will never return null, but will throw an exception when a
stream can't be opened.  This is unlikely to cause any problems because the routine
has always been able to throw exceptions, so all calling classes have already had to
include catch() blocks.  Throwing the exception rather than returning null is better
because it allows the reason for the exception to be given.

I also cleaned-up some of the status message so that they are more accurate.


git-svn-id: http://svn.code.sf.net/p/genoviz/code/trunk@765 80d5a9c2-7d8c-4c9f-9369-4e03bca4d2bd