Shortened the text variables used as messages for the HttpServletResponse.sendError() calls and removed newlines from these messages. Variables:        SERVER_SYNTAX_EXPLANATION        LIMITED_FEATURE_QUERIES_EXPLANATION
The aim here is to avoid tripping up Apache, which is not displaying
these messages but rather 502 'Bad Gateway' errors instead. My theory
is that these long messages are leading to invalid headers in the http
responses. This remains to be tested.

The message argument for HttpServletResponse.setError() is added
to the http header as well as to the TITLE and BODY elements in the
html page for the error. The interface doesn't provide the ability to
specify a shorter message for the header/title and a longer one for
the body, which would be nice.

A partial workaround to this error has been to add
'ProxyBadHeader Ignore' to the httpd.conf to tell Apache to ignore bad
http headers. Doing this provided a way to see the content of part of the
http header and gave a clue about the long message being the likely
culprit.

Saved the long versions of these as new variables, currently
unused. Plan to add this text at http://netaffxdas.affymetrix.com/das2,
some of which is already there.


git-svn-id: http://svn.code.sf.net/p/genoviz/code/trunk@3216 80d5a9c2-7d8c-4c9f-9369-4e03bca4d2bd