Move SimpleDas2Feature and Das2FeatureSaxParser from igb into genometryImpl.
This removes the final dependency between igb and das2_server, i.e.,
das2_server can now be built without first building igb.

It also puts the Das2 parser next to the Das1 parser.

The eventual goal would be to move most of the das1 and das2 code out of 
igb.



git-svn-id: http://svn.code.sf.net/p/genoviz/code/trunk@3406 80d5a9c2-7d8c-4c9f-9369-4e03bca4d2bd