Unfortunately, we can't really combine the tree view and the feature table due to Java limitations.  Neither the old Swing workarounds or the new Netbeans TreeTable can accomplish the other goals (namely, having an editable, always-in-focus combo box in the table).
Thus, we're relying on the old CheckTreeManager hack.  This is not ideal 
from a UI perspective, because it relies on extra clicks, and it's not 
intuitive that clicking a check box will add a feature to a list (of 
not-yet-fully-selected features).  However, that's what we're stuck with 
for now.

Other UI changes:

Move sequence list to right-hand side.  Since this is, in general, the 
last item selected, this is more correct from a UI perspective.

Change labels and defaults for combo boxes.  Clearer, uses less space.

Allow longer names in feature tree view.



git-svn-id: http://svn.code.sf.net/p/genoviz/code/trunk@3661 80d5a9c2-7d8c-4c9f-9369-4e03bca4d2bd