When working with booleans: Replace x |= blah (bitwise or) with x = x || blah (logical or). Replace y &= blah (bitwise and) with y = y && blah (logical and).
Due to short-circuiting, this allows the compiler to not evaluate the 
second argument.

SeqMapView at line 326 actually uses the evaluation of the second 
argument as a side-effect, so I left that.

Most of these locations also appear to use the same code due to 
copy-and-paste.  Should be fixed in the future.



git-svn-id: http://svn.code.sf.net/p/genoviz/code/trunk@3785 80d5a9c2-7d8c-4c9f-9369-4e03bca4d2bd