CompositeGraphSym: make code clearer.
GraphIntervalSym: Index and buffer the width coordinates. (To do: same with x and y.)
When a GraphIntervalSym is created, the width coordinates are written to a temp file and
read from there when needed (with a small buffer to increase performance).
This reduces memory requirements for loaded WIG files by 33%, with no noticeable decrease
in performance.



git-svn-id: http://svn.code.sf.net/p/genoviz/code/trunk@4610 80d5a9c2-7d8c-4c9f-9369-4e03bca4d2bd