Fixed an old bug in LocalUrlCacher (ReadContentIntoChunks) that was causing slowdowns and some large files (BNIBs) not to load correctly.
There's a lot of code dedicated to reading content in chunks when a header
is not detected.  However, this data was assumed to come in 64K chunks.  When
it came in smaller chunks, the memory was still being allocated... for some
Affy chromosomes (chr20 or higher in human), this was causing a 5000% increase
in memory usage, which was silently crashing the chromosome load.



git-svn-id: http://svn.code.sf.net/p/genoviz/code/trunk@4767 80d5a9c2-7d8c-4c9f-9369-4e03bca4d2bd