Due to some old design decisions, file/QuickLoad loading of graphs encounters a lot of problems when partially loading multiple tracks, or applying styles to tracks.
Here we explicitly force a unique track-level ID for GraphSyms that are loaded over file/QuickLoad,
which fixes the multiple tracks and style problems.

However, due to the old design decision, this adds a new potential problem for graph partial loading...
old IDs are generated uniquely, but not reproducibly.  This means pressing refresh multiple times
will add new tracks.

This is fixed for Gr,Sgr,and Wiggle below.  Other graph formats are special-cased not to allow
partial loading; will need to be revisited after the release.




git-svn-id: http://svn.code.sf.net/p/genoviz/code/trunk@6432 80d5a9c2-7d8c-4c9f-9369-4e03bca4d2bd