Bugfix: Search View was not returning all queries from DAS/2.
Turns out this was an old, known problem which may be due to the DAS2XML spec
and concepts like "ontology".
Multiple features with the same ID are being returned by the server, and
the client very specifically assumes this is a problem.

Fix (now possible, since all of the DAS/2 servers are of a recent vintage):
Request feature query information in BED format.  
Then ID collisions don't matter.



git-svn-id: http://svn.code.sf.net/p/genoviz/code/trunk@6980 80d5a9c2-7d8c-4c9f-9369-4e03bca4d2bd