Adding new felix.jar to repo.
Got it from:

http://mirrors.gigenet.com/apache//felix/org.apache.felix.main.distribution-4.2.1.tar.gz

12/19/2013

The felix.jar under ext (top level of project) is somemthing custom-built (?) HV 
obtained from a maven repository. It lacks a Main class and cannot be used to run
Lucene indexing here. To use that one would require some refactoring and adding a Main
class to this project.



git-svn-id: http://svn.code.sf.net/p/genoviz/code/trunk@15973 80d5a9c2-7d8c-4c9f-9369-4e03bca4d2bd